+++ src/main/java/roomescape/controller/reservation/ReservationThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.ReservationThemeRequest;
 import roomescape.dto.response.ReservationThemeResponse;
-import roomescape.service.ReservationThemeService;
+import roomescape.service.reservation.ReservationThemeService;
 
 @RequiredArgsConstructor
 @RequestMapping("/themes")
