+++ src/main/java/roomescape/domain/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.member;
 
 import java.util.List;
 import java.util.Optional;
