+++ src/main/java/roomescape/domain/reservationitem/ReservationItemRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservationitem;
 
 import java.time.LocalDate;
 import java.util.Optional;
