+++ src/main/java/roomescape/domain/reservationitem/ReservationThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservationitem;
 
 import java.time.LocalDate;
 import java.util.List;
