+++ src/main/java/roomescape/dto/response/MemberRegisterResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.response;
 
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 
 public record MemberRegisterResponse(
         long id,
