+++ src/main/java/roomescape/repository/impl/MemberRepositoryImpl.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Member;
-import roomescape.domain.MemberRepository;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
 import roomescape.repository.jpa.MemberJpaRepository;
 
 @RequiredArgsConstructor
