+++ src/main/java/roomescape/repository/impl/ReservationTimeRepositoryImpl.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.reservationitem.ReservationTime;
+import roomescape.domain.reservationitem.ReservationTimeRepository;
 import roomescape.repository.jpa.ReservationTimeJpaRepository;
 
 @RequiredArgsConstructor
