+++ src/main/java/roomescape/repository/jpa/MemberJpaRepository.java
@@ -3,7 +3,7 @@ package roomescape.repository.jpa;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 
 @Repository
 public interface MemberJpaRepository extends JpaRepository<Member, Long> {
