+++ src/main/java/roomescape/repository/jpa/ReservationThemeJpaRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTheme;
+import roomescape.domain.reservationitem.ReservationTheme;
 
 @Repository
 public interface ReservationThemeJpaRepository extends JpaRepository<ReservationTheme, Long> {
