+++ src/main/java/roomescape/repository/jpa/ReservationTimeJpaRepository.java
@@ -3,7 +3,7 @@ package roomescape.repository.jpa;
 import java.time.LocalTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservationitem.ReservationTime;
 
 @Repository
 public interface ReservationTimeJpaRepository extends JpaRepository<ReservationTime, Long> {
