+++ src/main/java/roomescape/service/member/MemberService.java
@@ -1,12 +1,12 @@
-package roomescape.service;
+package roomescape.service.member;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Member;
-import roomescape.domain.MemberRepository;
-import roomescape.domain.MemberRole;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
+import roomescape.domain.member.MemberRole;
 import roomescape.dto.request.MemberRegisterRequest;
 import roomescape.dto.response.MemberRegisterResponse;
 import roomescape.dto.response.MemberResponse;
