+++ src/main/java/roomescape/service/reservation/ReservationItemService.java
@@ -1,13 +1,13 @@
-package roomescape.service;
+package roomescape.service.reservation;
 
 import java.time.LocalDate;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.domain.ReservationItem;
-import roomescape.domain.ReservationItemRepository;
-import roomescape.domain.ReservationTheme;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservationitem.ReservationItem;
+import roomescape.domain.reservationitem.ReservationItemRepository;
+import roomescape.domain.reservationitem.ReservationTheme;
+import roomescape.domain.reservationitem.ReservationTime;
 
 @RequiredArgsConstructor
 @Service
