+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservationitem.ReservationTime;
 
 class ReservationTimeTest {
