+++ src/test/java/roomescape/repository/ReservationThemeRepositoryImplTest.java
@@ -16,15 +16,15 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestPropertySource;
 import roomescape.config.JpaConfig;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationItem;
-import roomescape.domain.ReservationItemRepository;
-import roomescape.domain.ReservationRepository;
-import roomescape.domain.ReservationStatus;
-import roomescape.domain.ReservationTheme;
-import roomescape.domain.ReservationThemeRepository;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservationitem.ReservationItem;
+import roomescape.domain.reservationitem.ReservationItemRepository;
+import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.reservation.ReservationStatus;
+import roomescape.domain.reservationitem.ReservationTheme;
+import roomescape.domain.reservationitem.ReservationThemeRepository;
+import roomescape.domain.reservationitem.ReservationTime;
+import roomescape.domain.reservationitem.ReservationTimeRepository;
 import roomescape.repository.impl.ReservationItemRepositoryImpl;
 import roomescape.repository.impl.ReservationRepositoryImpl;
 import roomescape.repository.impl.ReservationThemeRepositoryImpl;
