+++ src/test/java/roomescape/service/ReservationItemServiceTest.java
@@ -15,12 +15,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.TestPropertySource;
-import roomescape.domain.ReservationItem;
-import roomescape.domain.ReservationItemRepository;
-import roomescape.domain.ReservationTheme;
-import roomescape.domain.ReservationThemeRepository;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.reservationitem.ReservationItem;
+import roomescape.domain.reservationitem.ReservationItemRepository;
+import roomescape.domain.reservationitem.ReservationTheme;
+import roomescape.domain.reservationitem.ReservationThemeRepository;
+import roomescape.domain.reservationitem.ReservationTime;
+import roomescape.domain.reservationitem.ReservationTimeRepository;
+import roomescape.service.reservation.ReservationItemService;
 
 @SpringBootTest
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
